move _get and _set from container to section
